refactor(editor): rename createWithInsertBreak.ts->behavior.action.insert-break.ts
